requirements.txt
attrs==24.2.0
aws-cdk-lib==2.160.0
aws-cdk.asset-awscli-v1==2.2.204
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.aws-lambda-python-alpha==2.160.0a0
aws-cdk.cloud-assembly-schema==38.0.1
boto3==1.35.26
botocore==1.35.26
cattrs==23.2.3
certifi==2024.8.30
charset-normalizer==3.3.2
constructs==10.3.0
idna==3.10
importlib_resources==6.4.5
jmespath==1.0.1
jsii==1.103.1
multipart==1.0.0
publication==0.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.10.2
six==1.16.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==1.26.14
